refactor - separate model
